[refactor] Move SampleHolder into structures.py, unchanged
Groundwork, and nothing else: no behaviour changes, no call site changes, and the four
class bodies are identical before and after.

`SystemSettings` is about to hold a sample holder, which it cannot do from where the
class lives now. `fibsem/microscopes/_stage.py` imports from `structures.py`, so a
`holders` field on `SystemSettings` closes an import loop. The two ways out are a
deferred import inside the function that needs it, or moving the class below the loop.
The class moves: a deferred import would leave the cycle in place and hide it behind
an import that has to stay where it is, which is a thing the next person has to
rediscover.

`SampleGrid`, `SlotCalibration` and `GridSlot` come too. `SampleHolder` is built out
of them and they are of a piece; splitting them across two modules to move the minimum
would be worse than moving the group.

`_stage.py` re-exports all four plus `GRID_RADIUS`, so the roughly twenty modules that
import them from there are untouched. `fibsem.structures.SampleHolder is
fibsem.microscopes._stage.SampleHolder`.

## How this was checked

By AST, not by reading the diff. The four `ClassDef` nodes were dumped from `_stage.py`
before the move and compared against the ones parsed out of `structures.py` after it:
identical, and none still defined in `_stage.py`. A move this size is exactly where a
line goes missing without the diff looking wrong.

Full non-UI suite: 1 failed, 4927 passed -- the same failure, on the same test, as
before the move, and it fails on stock `origin/main` too.
