+++ fibsem/microscopes/_stage.py
@@ -16,363 +16,30 @@ from fibsem.config import (
     SAMPLE_HOLDER_CONFIGURATION_PATH,
     SAMPLE_HOLDER_OCCUPANCY_PATH,
 )
-from fibsem.structures import BeamType, FibsemStagePosition, RangeLimit
+from fibsem.structures import (
+    GRID_RADIUS,
+    BeamType,
+    FibsemStagePosition,
+    GridSlot,
+    RangeLimit,
+    SampleGrid,
+    SampleHolder,
+    SlotCalibration,
+)
+
+# Re-exported: these four moved into `structures.py` so `SystemSettings` could hold a
+# holder without closing an import loop. They are still part of this module's surface
+# -- around twenty call sites import them from here.
+__all__ = [
+    "GRID_RADIUS",
+    "SampleGrid",
+    "SlotCalibration",
+    "GridSlot",
+    "SampleHolder",
+]
 
 if TYPE_CHECKING:
     from fibsem.microscope import FibsemMicroscope
-
-GRID_RADIUS = 1e-3  # 1mm
-
-
-@dataclass
-class SampleGrid:
-    """A physical TEM grid or sample that can be loaded into a GridSlot."""
-
-    name: str
-    description: str = ""
-    radius: float = field(
-        default=GRID_RADIUS,
-        metadata={"unit": "mm", "tooltip": "Radius of the sample grid", "scale": 1e3},
-    )
-
-    def to_dict(self) -> dict:
-        return {
-            "name": self.name,
-            "description": self.description,
-            "radius": self.radius,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "SampleGrid":
-        return SampleGrid(
-            name=data.get("name", ""),
-            description=data.get("description", ""),
-            radius=data.get("radius", GRID_RADIUS),
-        )
-
-
-@dataclass
-class SlotCalibration:
-    """The proof that a slot position was captured properly, and what it was captured against.
-
-    Written only by the calibration wizard. A position without one, which is every
-    holder file in the field before this existed, is not trusted: it was captured at
-    an unknown orientation with a button that took whatever the stage said. A position
-    whose ``pre_tilt`` or ``rotation_reference`` no longer match the system
-    configuration is not trusted either, since the geometry it was captured against
-    has moved. Both cases read as "not calibrated" and the wizard is the way back.
-    """
-
-    orientation: str
-    pre_tilt: float
-    rotation_reference: float
-    captured_at: str = ""
-    fibsem_version: str = ""
-
-    @classmethod
-    def builtin(cls, pre_tilt: float, rotation_reference: float) -> "SlotCalibration":
-        """A position the hardware defines, not one an operator captured.
-
-        The compustage working slot is at the compustage origin by construction:
-        the autoloader puts every grid at the same place and the coordinate system
-        is referenced to it. There is nothing to capture, so the record says so.
-        """
-        return cls(
-            orientation="SEM",
-            pre_tilt=pre_tilt,
-            rotation_reference=rotation_reference,
-            captured_at="",
-            fibsem_version="built-in",
-        )
-
-    @property
-    def is_builtin(self) -> bool:
-        return self.fibsem_version == "built-in" and not self.captured_at
-
-    def matches(self, pre_tilt: float, rotation_reference: float) -> bool:
-        return (
-            abs(self.pre_tilt - pre_tilt) < 1e-3
-            and abs(self.rotation_reference - rotation_reference) < 1e-3
-        )
-
-    def to_dict(self) -> dict:
-        return {
-            "orientation": self.orientation,
-            "pre_tilt": self.pre_tilt,
-            "rotation_reference": self.rotation_reference,
-            "captured_at": self.captured_at,
-            "fibsem_version": self.fibsem_version,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "SlotCalibration":
-        return SlotCalibration(
-            orientation=str(data.get("orientation", "")),
-            pre_tilt=float(data.get("pre_tilt", 0.0)),
-            rotation_reference=float(data.get("rotation_reference", 0.0)),
-            captured_at=str(data.get("captured_at", "")),
-            fibsem_version=str(data.get("fibsem_version", "")),
-        )
-
-
-@dataclass
-class GridSlot:
-    """A slot that may hold one SampleGrid.
-
-    A holder *working* slot has a stage ``position`` once it has been calibrated, and
-    a ``calibration`` record saying so; until then ``position`` is None and nothing
-    will move to it. A loader *magazine* slot is storage and never has a position.
-    """
-
-    name: str
-    index: int
-    position: Optional[FibsemStagePosition] = None
-    loaded_grid: Optional[SampleGrid] = None
-    calibration: Optional[SlotCalibration] = None
-
-    @property
-    def is_calibrated(self) -> bool:
-        return self.position is not None and self.calibration is not None
-
-    def to_dict(self) -> dict:
-        return {
-            "name": self.name,
-            "index": self.index,
-            "position": self.position.to_dict() if self.position is not None else None,
-            "loaded_grid": self.loaded_grid.to_dict()
-            if self.loaded_grid is not None
-            else None,
-            "calibration": self.calibration.to_dict()
-            if self.calibration is not None
-            else None,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "GridSlot":
-        loaded_grid_data = data.get("loaded_grid")
-        loaded_grid = (
-            SampleGrid.from_dict(loaded_grid_data)
-            if loaded_grid_data is not None
-            else None
-        )
-        position_data = data.get("position")
-        position = (
-            FibsemStagePosition(**position_data) if position_data is not None else None
-        )
-        calibration_data = data.get("calibration")
-        calibration = (
-            SlotCalibration.from_dict(calibration_data)
-            if calibration_data is not None
-            else None
-        )
-        slot = GridSlot(
-            name=data.get("name", ""),
-            index=data.get("index", 0),
-            position=position,
-            loaded_grid=loaded_grid,
-            calibration=calibration,
-        )
-        if slot.position is not None:
-            slot.position.name = slot.name
-        return slot
-
-
-@dataclass
-class SampleHolder:
-    name: str = field(
-        default="Sample Holder", metadata={"tooltip": "Name of the sample holder"}
-    )
-    description: str = field(
-        default="", metadata={"tooltip": "Description of the sample holder"}
-    )
-    capacity: int = field(
-        default=2,
-        metadata={
-            "minimum": 1,
-            "maximum": 12,
-            "tooltip": "Number of grid slots on this holder",
-        },
-    )
-    slots: dict[str, GridSlot] = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        self._parent: Optional["FibsemMicroscope"] = None
-
-    @property
-    def pre_tilt(self) -> float:
-        if self._parent is not None:
-            return self._parent.system.stage.shuttle_pre_tilt
-        return 0.0
-
-    @property
-    def reference_rotation(self) -> float:
-        if self._parent is not None:
-            return self._parent.system.stage.rotation_reference
-        return 0.0
-
-    def find_slot_for_grid(self, grid: "SampleGrid") -> Optional["GridSlot"]:
-        """Return the slot that has this SampleGrid loaded, or None."""
-        for slot in self.slots.values():
-            if slot.loaded_grid is not None and slot.loaded_grid.name == grid.name:
-                return slot
-        return None
-
-    def find_slot_by_grid_name(self, grid_name: str) -> Optional["GridSlot"]:
-        """Return the slot whose loaded grid matches the given name, or None."""
-        for slot in self.slots.values():
-            if slot.loaded_grid is not None and slot.loaded_grid.name == grid_name:
-                return slot
-        return None
-
-    @property
-    def occupied_slots(self) -> List["GridSlot"]:
-        """The working slots that hold a grid: what is loaded right now."""
-        return [
-            s
-            for s in sorted(self.slots.values(), key=lambda s: s.index)
-            if s.loaded_grid is not None
-        ]
-
-    @property
-    def calibrated_slots(self) -> List["GridSlot"]:
-        """The slots with a trusted position: the ones the stage can be sent to."""
-        return [
-            s
-            for s in sorted(self.slots.values(), key=lambda s: s.index)
-            if s.is_calibrated
-        ]
-
-    def discard_untrusted_positions(
-        self, pre_tilt: float, rotation_reference: float
-    ) -> List[str]:
-        """Drop every slot position that was not calibrated against this geometry.
-
-        A position with no calibration record was captured by the old per-slot
-        button at an unknown orientation; one whose record disagrees with the current
-        pre-tilt or reference rotation was captured against a stage that has since
-        been reconfigured. Neither can be trusted, so both become "not calibrated"
-        in memory. The file is left alone: the wizard rewrites it when someone
-        recalibrates, and never before. Returns one line per discarded slot, for
-        the log and the UI.
-        """
-        notes: List[str] = []
-        for slot in sorted(self.slots.values(), key=lambda s: s.index):
-            if slot.position is None:
-                continue
-            if slot.calibration is None:
-                reason = "it has no calibration record"
-            elif not slot.calibration.matches(pre_tilt, rotation_reference):
-                reason = (
-                    f"it was calibrated at pre-tilt {slot.calibration.pre_tilt:g}°, "
-                    f"reference rotation {slot.calibration.rotation_reference:g}°, "
-                    f"but the system is configured for {pre_tilt:g}° / "
-                    f"{rotation_reference:g}°"
-                )
-            else:
-                continue
-            slot.position = None
-            slot.calibration = None
-            notes.append(f"{slot.name}: position discarded because {reason}")
-        return notes
-
-    def _ensure_slots(self) -> None:
-        """Ensure exactly `capacity` slots exist; add empty ones for missing indices."""
-        for i in range(self.capacity):
-            name = f"Slot-{i + 1:02d}"
-            if name not in self.slots:
-                # A new slot has no position until it is calibrated; inventing one
-                # at the origin was a number that looked like a measurement.
-                self.slots[name] = GridSlot(name=name, index=i, position=None)
-        for name in [
-            n for n, s in list(self.slots.items()) if s.index >= self.capacity
-        ]:
-            del self.slots[name]
-
-    def to_dict(self, include_grids: bool = True) -> dict:
-        slots = {}
-        for name, slot in self.slots.items():
-            data = slot.to_dict()
-            if not include_grids:
-                data["loaded_grid"] = None
-            slots[name] = data
-        return {
-            "name": self.name,
-            "capacity": self.capacity,
-            "slots": slots,
-            "description": self.description,
-        }
-
-    # -- occupancy: which grid is in which slot, kept apart from the calibration --
-
-    def occupancy_to_dict(self) -> dict:
-        """Slot name -> grid, for the slots that hold one."""
-        return {
-            name: slot.loaded_grid.to_dict()
-            for name, slot in self.slots.items()
-            if slot.loaded_grid is not None
-        }
-
-    def apply_occupancy(self, data: dict) -> None:
-        """Put the recorded grids back into their slots; unlisted slots are emptied."""
-        for name, slot in self.slots.items():
-            grid_data = (data or {}).get(name)
-            slot.loaded_grid = (
-                SampleGrid.from_dict(grid_data) if grid_data is not None else None
-            )
-
-    def save_occupancy(self, path: Union[str, Path]) -> None:
-        path = Path(path)
-        path.parent.mkdir(parents=True, exist_ok=True)
-        with open(path, "w") as f:
-            yaml.dump(
-                self.occupancy_to_dict(), f, default_flow_style=False, sort_keys=False
-            )
-
-    def load_occupancy(self, path: Union[str, Path]) -> bool:
-        """Apply the occupancy file if there is one. Returns whether there was."""
-        path = Path(path)
-        if not path.exists():
-            return False
-        with open(path, "r") as f:
-            self.apply_occupancy(yaml.safe_load(f) or {})
-        return True
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "SampleHolder":
-        slots = {
-            name: GridSlot.from_dict(slot_data)
-            for name, slot_data in data.get("slots", {}).items()
-        }
-        holder = SampleHolder(
-            name=data.get("name", "Sample Holder"),
-            capacity=data.get("capacity", max(len(slots), 1)),
-            slots=slots,
-            description=data.get("description", ""),
-        )
-        holder._ensure_slots()
-        return holder
-
-    @classmethod
-    def load(cls, path: Union[str, Path]) -> "SampleHolder":
-        path = Path(path)
-        if not path.exists():
-            raise FileNotFoundError(f"Sample holder config not found: {path}")
-        with open(path, "r") as f:
-            data = yaml.safe_load(f)
-        return cls.from_dict(data)
-
-    def save(self, path: Union[str, Path]) -> None:
-        """Write the holder's geometry and calibration. Not the grids in it: those
-        are session state and live in the occupancy file (``save_occupancy``)."""
-        path = Path(path)
-        path.parent.mkdir(parents=True, exist_ok=True)
-        with open(path, "w") as f:
-            yaml.dump(
-                self.to_dict(include_grids=False),
-                f,
-                default_flow_style=False,
-                sort_keys=False,
-            )
 
 
 class GridExchangeError(RuntimeError):
